+++ cosmohub/api/rest/acls.py
@@ -103,26 +103,26 @@ class AclItem(Resource):
 
         attrs = parser.parse_args(strict=True)
         
+        # TODO: assert set(granted) disjoint set(revoked)
+        
         with transactional_session(db.session) as session:
             groups_granted = session.query(
-                model.User
-            ).filter(
-                model.User.id == g.session['user'].id
+                model.Group
             ).join(
-                model.User.groups_administered,
+                model.Group.users_admins, # @UndefinedVariable
             ).filter(
+                model.User.id == g.session['user'].id,
                 model.Group.name.in_(attrs['groups_granted']),
-            ).with_for_update().one().groups_administered
+            ).with_for_update().all()
             
             groups_revoked = session.query(
-                model.User
-            ).filter(
-                model.User.id == g.session['user'].id
+                model.Group
             ).join(
-                model.User.groups_administered,
+                model.Group.users_admins, # @UndefinedVariable
             ).filter(
+                model.User.id == g.session['user'].id,
                 model.Group.name.in_(attrs['groups_revoked']),
-            ).with_for_update().one().groups_administered
+            ).with_for_update().all()
             
             user = session.query(
                 model.User
@@ -142,28 +142,31 @@ class AclItem(Resource):
             for group in groups_granted:
                 if group not in user.acls:
                     user.acls[group] = model.ACL(group=group, user=user)
-                user.acls[group].is_granted = True
+                if user.acls[group].is_granted != True:
+                    user.acls[group].is_granted = True
             
             for group in groups_revoked:
                 if group not in user.acls:
                     user.acls[group] = model.ACL(group=group, user=user)
-                user.acls[group].is_granted = False
+                if user.acls[group].is_granted != False:
+                    user.acls[group].is_granted = False
             
-            session.flush()
+            if session.dirty:
+                session.flush()
             
-            if attrs['notify'] and session.dirty:
-                groups = [
-                    group
-                    for group in user.acls
-                    if user.acls[group].is_granted
-                ]
-                
-                mail.send_message(
-                    subject = current_app.config['MAIL_SUBJECTS']['acls_updated'],
-                    recipients = [user.email],
-                    body = render_template('mail/acls_updated.txt', groups=groups),
-                    html = render_template('mail/acls_updated.html', groups=groups),
-                )
+                if attrs['notify']:
+                    groups = [
+                        group
+                        for group in user.acls
+                        if user.acls[group].is_granted
+                    ]
+                    
+                    mail.send_message(
+                        subject = current_app.config['MAIL_SUBJECTS']['acls_updated'],
+                        recipients = [user.email],
+                        body = render_template('mail/acls_updated.txt', user=user, groups=groups),
+                        html = render_template('mail/acls_updated.html', user=user, groups=groups),
+                    )
             
             g.session['track']({
                 't' : 'event',